requirements.txt
pyyaml>=6.0.1
pathspec>=0.12.1
mcp>=1.0.0